Adapted Function attachEventListeners so it still works with new chat logic
