+++ frontend/index.html
@@ -11,11 +11,13 @@
 </head>
 
 <body class="bg-stone-900">
+	<!-- Chat Button -->
 	<button id="chatButton" style="position: absolute; right: 2%; bottom: 2%;" class="p-2 text-gray-900 rounded-xl dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700">
 		<svg class="w-6 h-6" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 24 24">
 			<path d="M20 2H4c-1.1 0-2 .9-2 2v18l4-4h14c1.1 0 2-.9 2-2V4c0-1.1-.9-2-2-2zm0 14H6l-2 2V4h16v12z"/>
 		</svg>
 	</button>
+
 	<div id="chatMenu" style="display:none; height: 400px; flex-direction: column; width: 1000px; position: absolute; bottom: 10px; right: 20px; z-index: 100; background-color: rgba(0, 0, 0, 0.9); color: white; padding: 20px; border-radius: 10px; text-align: center;">
         <div class="flex justify-between items-center mb-4">
             <h1>CHAT</h1>
@@ -26,7 +28,10 @@
             </button>
 			
         </div>
+		<!-- Users List -->
 		<div class="users-list scroll-y-auto overflow-y-auto max-h-96"></div>
+
+		<!-- Chat Discussion -->
         <div id="chatScroll" class="flex items-start flex-col" style="flex-grow: 1; overflow-y: auto;">
           
         </div>
